pyproject.toml
[tool.poetry]
name = "incident-intelligence-engine"
version = "0.1.0"
description = "Transforms unstructured incident reports into actionable intelligence using LLMs and n8n automation."
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]
pythonpath = ["src"]

[project]
name = "incident-intelligence-engine"
version = "0.1.0"
description = "Transforms unstructured incident reports into actionable intelligence using LLMs and n8n automation."
authors = [
	{ name = "Your Name", email = "your.email@example.com" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
	"fastapi",
	"uvicorn",
	"pydantic",
	"httpx",
	"openai",
	"langdetect",
	"SQLAlchemy",
	"alembic",
	"python-dotenv",
	"loguru",
	"pytest",
	"pytest-asyncio",
	"requests",
	"PyYAML",
	"jsonschema"
]
classifiers = [
	"Programming Language :: Python :: 3",
	"License :: OSI Approved :: MIT License",
	"Operating System :: OS Independent"
]
packages = [{ include = "src" }]
dynamic = ["version"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
pydantic = "^2.6.0"
httpx = "^0.27.0"
openai = "^1.13.3"
langdetect = "^1.0.9"
SQLAlchemy = "^2.0.29"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5"
requests = "^2.31.0"
PyYAML = "^6.0.1"
jsonschema = "^4.21.1"

[tool.poetry.dev-dependencies]
black = "^24.3.0"
flake8 = "^7.0.0"
isort = "^5.13.2"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
